feat: add tweet stats

+++ scraper/scrape.py
@@ -3,7 +3,8 @@ import twint
 
 OUTPUT_FILE = f"{os.path.dirname(__file__)}/output.csv"
 
-os.remove(OUTPUT_FILE)
+if os.path.exists(OUTPUT_FILE):
+    os.remove(OUTPUT_FILE)
 
 c = twint.Config()
 
@@ -13,7 +14,7 @@ c.Search = "zig"
 c.Stats = True
 c.Hide_output = True
 c.Store_csv = True
-c.Custom["tweet"] = ["id", "username"]
+c.Custom["tweet"] = ["id", "username", "likes_count", "replies_count", "retweets_count"]
 c.Output = OUTPUT_FILE
 
 twint.run.Search(c)
